Removing wrong encodeURI for thumbnail URLs in the admin interface
The encodeURI is only working by accident. The image URLs in ezp do not contain any characters that need encoding. So the output with or without encodeURI is always the same. It's causing an issue with the image server, because we allow other characters like Umlaute etc. The extra encodeURI is breaking the URL to the image.
eZ is using `eZURI::transformURI( $thumbUrl, true, null, false );` to build the thumbnail URL. I don't see why we'd need encodeURI before we set an image src tag.